+++ R/integration.R
@@ -1041,7 +1041,7 @@ FindTransferAnchors <- function(
         reduction = query[[reference.reduction]],
         data = GetAssayData(object = reference, assay = reference.assay, slot = "data"),
         mode = "lsi",
-        do.center = TRUE,
+        do.center = FALSE,
         do.scale = FALSE,
         use.original.stats = FALSE,
         verbose = verbose
@@ -1053,7 +1053,7 @@ FindTransferAnchors <- function(
         reduction = reference[[reference.reduction]],
         data = GetAssayData(object = query, assay = query.assay, slot = "data"),
         mode = "lsi",
-        do.center = TRUE,
+        do.center = FALSE,
         do.scale = FALSE,
         use.original.stats = FALSE,
         verbose = verbose
@@ -1271,8 +1271,23 @@ GetTransferPredictions <- function(object, assay = "predictions", slot = "data",
 #' query, as the merged object will subsequently contain more cells than was in
 #' query, and weights will need to be calculated for all cells in the object.
 #' @param sd.weight Controls the bandwidth of the Gaussian kernel for weighting
-#' @param sample.tree Specify the order of integration. If NULL, will compute
-#' automatically.
+#' @param sample.tree Specify the order of integration. Order of integration
+#' should be encoded in a matrix, where each row represents one of the pairwise
+#' integration steps. Negative numbers specify a dataset, positive numbers
+#' specify the integration results from a given row (the format of the merge
+#' matrix included in the \code{\link{hclust}} function output). For example:
+#' `matrix(c(-2, 1, -3, -1), ncol = 2)` gives:
+#'
+#' ```
+#'             [,1]  [,2]
+#'        [1,]   -2   -3
+#'        [2,]    1   -1
+#' ```
+#'
+#' Which would cause dataset 2 and 3 to be integrated first, then the resulting
+#' object integrated with dataset 1.
+#'
+#'  If NULL, the sample tree will be computed automatically.
 #' @param preserve.order Do not reorder objects based on size for each pairwise
 #' integration.
 #' @param eps Error bound on the neighbor finding algorithm (from
@@ -1291,6 +1306,7 @@ GetTransferPredictions <- function(object, assay = "predictions", slot = "data",
 #'
 #' @export
 #' @concept integration
+#' @md
 #' @examples
 #' \dontrun{
 #' # to install the SeuratData package see https://github.com/satijalab/seurat-data