+++ R/visualization.R
@@ -555,6 +555,9 @@ RidgePlot <- function(
 #' single violin shapes.
 #' @param adjust Adjust parameter for geom_violin
 #' @param flip flip plot orientation (identities on x-axis)
+#' @param raster Convert points to raster format. Requires 'ggrastr' to be installed.
+# default is \code{NULL} which automatically rasterizes if ggrastr is installed and
+# number of points exceed 100,000.
 #'
 #' @return A \code{\link[patchwork]{patchwork}ed} ggplot object if
 #' \code{combine = TRUE}; otherwise, a list of ggplot objects
@@ -589,7 +592,8 @@ VlnPlot <- function(
   stack = FALSE,
   combine = TRUE,
   fill.by = 'feature',
-  flip = FALSE
+  flip = FALSE,
+  raster = NULL
 ) {
   if (
     !is.null(x = split.by) &
@@ -624,7 +628,8 @@ VlnPlot <- function(
     stack = stack,
     combine = combine,
     fill.by = fill.by,
-    flip = flip
+    flip = flip,
+    raster = raster
   ))
 }
 
@@ -652,11 +657,12 @@ VlnPlot <- function(
 #' @seealso \code{\link{DimPlot}}
 #'
 #' @examples
-#' data("pbmc_small")
-#' pbmc_small
-#' pbmc_small <- BuildClusterTree(object = pbmc_small, verbose = FALSE)
-#' PlotClusterTree(pbmc_small)
-#' ColorDimSplit(pbmc_small, node = 5)
+#' if (requireNamespace("ape", quietly = TRUE)) {
+#'   data("pbmc_small")
+#'   pbmc_small <- BuildClusterTree(object = pbmc_small, verbose = FALSE)
+#'   PlotClusterTree(pbmc_small)
+#'   ColorDimSplit(pbmc_small, node = 5)
+#' }
 #'
 ColorDimSplit <- function(
   object,
@@ -981,6 +987,7 @@ FeaturePlot <- function(
   blend.threshold = 0.5,
   label = FALSE,
   label.size = 4,
+  label.color = "black",
   repel = FALSE,
   ncol = NULL,
   coord.fixed = FALSE,
@@ -1263,7 +1270,8 @@ FeaturePlot <- function(
           plot = plot,
           id = 'ident',
           repel = repel,
-          size = label.size
+          size = label.size,
+          color = label.color
         )
       }
       # Make FeatureHeatmaps look nice(ish)
@@ -3824,10 +3832,11 @@ JackStrawPlot <- function(
 #' @concept visualization
 #'
 #' @examples
-#' data("pbmc_small")
-#' pbmc_small <- BuildClusterTree(object = pbmc_small)
-#' PlotClusterTree(object = pbmc_small)
-#'
+#' if (requireNamespace("ape", quietly = TRUE)) {
+#'   data("pbmc_small")
+#'   pbmc_small <- BuildClusterTree(object = pbmc_small)
+#'   PlotClusterTree(object = pbmc_small)
+#' }
 PlotClusterTree <- function(object, direction = "downwards", ...) {
   if (!PackageCheck('ape', error = FALSE)) {
     stop(cluster.ape, call. = FALSE)
@@ -5555,6 +5564,8 @@ Col2Hex <- function(...) {
 # ggplot object. If \code{FALSE}, return a list of ggplot objects
 # @param fill.by Color violins/ridges based on either 'feature' or 'ident'
 # @param flip flip plot orientation (identities on x-axis)
+# @param raster Convert points to raster format, default is \code{NULL} which
+# automatically rasterizes if plotting more than 100,000 cells
 #
 # @return A \code{\link[patchwork]{patchwork}ed} ggplot object if
 # \code{combine = TRUE}; otherwise, a list of ggplot objects
@@ -5583,7 +5594,8 @@ ExIPlot <- function(
   stack = FALSE,
   combine = TRUE,
   fill.by = NULL,
-  flip = FALSE
+  flip = FALSE,
+  raster = NULL
 ) {
   assay <- assay %||% DefaultAssay(object = object)
   DefaultAssay(object = object) <- assay
@@ -5684,7 +5696,8 @@ ExIPlot <- function(
         adjust = adjust,
         cols = cols,
         pt.size = pt.size,
-        log = log
+        log = log,
+        raster = raster
       ))
     }
   )
@@ -7188,6 +7201,9 @@ SingleDimPlot <- function(
 #' @param cols Colors to use for plotting
 #' @param seed.use Random seed to use. If NULL, don't set a seed
 #' @param log plot Y axis on log scale
+#' @param raster Convert points to raster format. Requires 'ggrastr' to be installed.
+#' default is \code{NULL} which automatically rasterizes if ggrastr is installed and
+#' number of points exceed 100,000.
 #'
 #' @return A ggplot-based Expression-by-Identity plot
 #'
@@ -7213,8 +7229,23 @@ SingleExIPlot <- function(
   pt.size = 0,
   cols = NULL,
   seed.use = 42,
-  log = FALSE
+  log = FALSE,
+  raster = NULL
 ) {
+   if (!is.null(x = raster) && isTRUE(x = raster)){
+    if (!PackageCheck('ggrastr', error = FALSE)) {
+      stop("Please install ggrastr from CRAN to enable rasterization.")
+    }
+  }
+  if (PackageCheck('ggrastr', error = FALSE)) {
+    # Set rasterization to true if ggrastr is installed and
+    # number of points exceeds 100,000
+    if ((nrow(x = data) > 1e5) & !isFALSE(raster)){
+      message("Rasterizing points since number of points exceeds 100,000.",
+              "\nTo disable this behavior set `raster=FALSE`")
+    }
+    raster <- TRUE
+  }
   if (!is.null(x = seed.use)) {
     set.seed(seed = seed.use)
   }
@@ -7274,13 +7305,25 @@ SingleExIPlot <- function(
         theme(axis.text.x = element_text(angle = 45, hjust = 1))
       )
       if (is.null(x = split)) {
-        jitter <- geom_jitter(height = 0, size = pt.size, show.legend = FALSE)
+        if (isTRUE(x = raster)) {
+          jitter <- ggrastr::rasterize(geom_jitter(height = 0, size = pt.size, show.legend = FALSE))
+        } else {
+          jitter <- geom_jitter(height = 0, size = pt.size, show.legend = FALSE)
+        }
       } else {
-        jitter <- geom_jitter(
-          position = position_jitterdodge(jitter.width = 0.4, dodge.width = 0.9),
-          size = pt.size,
-          show.legend = FALSE
-        )
+        if (isTRUE(x = raster)) {
+          jitter <- ggrastr::rasterize(geom_jitter(
+            position = position_jitterdodge(jitter.width = 0.4, dodge.width = 0.9),
+            size = pt.size,
+            show.legend = FALSE
+          ))
+        } else {
+          jitter <- geom_jitter(
+            position = position_jitterdodge(jitter.width = 0.4, dodge.width = 0.9),
+            size = pt.size,
+            show.legend = FALSE
+          )
+        }
       }
       log.scale <- scale_y_log10()
       axis.scale <- ylim